Move the core of validateAsync() to NewDirectoryPage class
